requirements.txt
celery==3.1.19
Django==1.9
psycopg2==2.6.1
selenium==2.32.0
PyYAML==3.11
django-pipeline==1.5.4